Split Ether.Internal
