+++ src/Ether/Internal/TupleInstances_HasLens.hs
@@ -1,0 +1,11 @@
+module Ether.Internal.TupleInstances_HasLens () where
+
+import Ether.Internal.TupleInstances_HasLens_1 ()
+import Ether.Internal.TupleInstances_HasLens_2 ()
+import Ether.Internal.TupleInstances_HasLens_3 ()
+import Ether.Internal.TupleInstances_HasLens_4 ()
+import Ether.Internal.TupleInstances_HasLens_5 ()
+import Ether.Internal.TupleInstances_HasLens_6 ()
+import Ether.Internal.TupleInstances_HasLens_7 ()
+import Ether.Internal.TupleInstances_HasLens_8 ()
+import Ether.Internal.TupleInstances_HasLens_9 ()
