+++ src/Ether/Internal/TupleInstances_HasLens_2.hs
@@ -1,0 +1,7 @@
+{-# OPTIONS -fno-warn-orphans #-}
+
+module Ether.Internal.TupleInstances_HasLens_2 () where
+
+import Ether.Internal.TH_TupleInstances (makeTupleInstancesHasLens)
+
+makeTupleInstancesHasLens [21..29]
