+++ src/Ether/Internal/TupleInstances_HasLens_3.hs
@@ -1,0 +1,7 @@
+{-# OPTIONS -fno-warn-orphans #-}
+
+module Ether.Internal.TupleInstances_HasLens_3 () where
+
+import Ether.Internal.TH_TupleInstances (makeTupleInstancesHasLens)
+
+makeTupleInstancesHasLens [30..36]
